protobuf/test_proto/requirements.txt
protobuf3-to-dict
pyyaml
